+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Processors and Interactions are now shown in a component-inspector-like fashion in the Input Action editor window, allowing you to see the properties of all items at once.
 - The various `InputAction.lastTriggerXXX` APIs have been removed.
   * Rationale: They have very limited usefulness and if you need the information, it's easy to set things up in order to keep track of it yourself. Also, we plan on having a polling API for actions in the future which is really what the `lastActionXXX` APIs were trying to (imperfectly) solve.
+- `Tap`, `SlowTap`, and `MultiTap` interactions now respect button press points.
 
 ### Fixed